+++ src/onegov.activity/onegov/activity/models/__init__.py
@@ -2,7 +2,8 @@ from onegov.activity.models.activity import Activity, ACTIVITY_STATES
 from onegov.activity.models.attendee import Attendee
 from onegov.activity.models.booking import Booking
 from onegov.activity.models.invoice_item import InvoiceItem
-from onegov.activity.models.occasion import Occasion, DAYS
+from onegov.activity.models.occasion import Occasion
+from onegov.activity.models.occasion_date import OccasionDate, DAYS
 from onegov.activity.models.period import Period
 
 
@@ -12,6 +13,7 @@ __all__ = [
     'Booking',
     'InvoiceItem',
     'Occasion',
+    'OccasionDate',
     'Period',
     'ACTIVITY_STATES',
     'DAYS'